feat(theme): fetch and merge theme from API
