+++ package.json
@@ -31,11 +31,12 @@
     "@types/react-scrollspy": "3.3.2",
     "@types/react-syntax-highlighter": "11.0.4",
     "@types/redux-logger": "3.0.7",
-    "axios": "^0.21.1",
+    "axios": "0.21.1",
     "classnames": "2.2.6",
     "clsx": "1.0.4",
     "connected-react-router": "6.6.1",
-    "formik": "^2.2.0",
+    "deepmerge": "4.2.2",
+    "formik": "2.2.0",
     "generate-password": "1.4.2",
     "history": "4.10.1",
     "i18next": "19.0.1",
@@ -43,16 +44,16 @@
     "js-cookie": "2.2.1",
     "material-ui-dots": "2.0.0",
     "react": "16.12.0",
-    "react-animated-slider": "^2.0.0",
-    "react-cookie-consent": "^6.1.0",
+    "react-animated-slider": "2.0.0",
+    "react-cookie-consent": "6.1.0",
     "react-dom": "16.12.0",
     "react-i18next": "11.2.5",
-    "react-material-ui-carousel": "^2.0.2",
+    "react-material-ui-carousel": "2.0.2",
     "react-redux": "7.1.3",
     "react-router": "5.1.2",
     "react-router-dom": "5.1.2",
     "react-scrollspy": "3.4.2",
-    "react-slidy": "^4.1.0",
+    "react-slidy": "4.1.0",
     "react-swipeable-views": "0.13.3",
     "react-swipeable-views-utils": "0.13.9",
     "react-syntax-highlighter": "12.2.1",
@@ -64,10 +65,10 @@
     "swagger-ui-react": "3.36.2",
     "tinycolor2": "1.4.1",
     "typeface-roboto": "0.0.75",
-    "yup": "^0.29.3"
+    "yup": "0.29.3"
   },
   "devDependencies": {
-    "@apisuite/extension-ui-types": "^1.0.5",
+    "@apisuite/extension-ui-types": "1.0.5",
     "@babel/cli": "7.7.4",
     "@babel/core": "7.7.4",
     "@babel/plugin-proposal-class-properties": "7.7.4",
@@ -80,6 +81,7 @@
     "@material-ui/core": "4.11.0",
     "@material-ui/icons": "4.5.1",
     "@material-ui/lab": "4.0.0-alpha.56",
+    "@semantic-release/git": "9.0.0",
     "@types/classnames": "2.2.9",
     "@types/i18n-js": "3.0.1",
     "@types/js-cookie": "2.2.4",
@@ -94,7 +96,6 @@
     "@types/webpack-env": "1.14.1",
     "@typescript-eslint/eslint-plugin": "2.20.0",
     "@typescript-eslint/parser": "2.20.0",
-    "@semantic-release/git": "^9.0.0",
     "autoprefixer": "9.7.3",
     "babel-eslint": "10.0.3",
     "babel-jest": "24.9.0",
@@ -102,7 +103,7 @@
     "babel-minify-webpack-plugin": "0.3.1",
     "babel-plugin-jsx-remove-data-test-id": "2.1.3",
     "commitizen": "4.2.1",
-    "core-js": "^3.4.7",
+    "core-js": "3.4.7",
     "cross-env": "6.0.3",
     "css-hot-loader": "1.4.4",
     "css-loader": "3.6.0",
@@ -138,8 +139,8 @@
     "redux-saga-test-plan": "4.0.0-rc.3",
     "rimraf": "3.0.0",
     "robotstxt-webpack-plugin": "7.0.0",
-    "semantic-release": "^17.3.8",
     "sass-loader": "8.0.0",
+    "semantic-release": "17.3.8",
     "serve": "11.3.2",
     "style-loader": "1.0.1",
     "typescript": "3.9.5",
